backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
celery==5.4.0
redis==5.0.4
pillow==10.3.0
qrcode==7.4.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
stripe==9.9.0
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.31.0
